DamarisVar: add translation unit
