+++ CMakeLists_files.cmake
@@ -623,6 +623,7 @@ if (Damaris_FOUND AND MPI_FOUND)
   list (APPEND PUBLIC_HEADER_FILES ebos/damariswriter.hh)
   list (APPEND PUBLIC_HEADER_FILES opm/simulators/utils/DamarisVar.hpp)
   list (APPEND PUBLIC_HEADER_FILES opm/simulators/utils/GridDataOutput.hpp)
+  list(APPEND MAIN_SOURCE_FILES opm/simulators/utils/DamarisVar.cpp)
 endif()
 
 if(HDF5_FOUND)
